Move adapter logic to strategy.
